+++ cpp/src/io/functions.cpp
@@ -151,6 +151,16 @@ table_with_metadata read_csv(read_csv_args const& args, rmm::mr::device_memory_r
   }
 }
 
+// Freeform API wraps the detail writer class API
+void write_csv(write_csv_args const& args, rmm::mr::device_memory_resource* mr)
+{
+  using namespace cudf::experimental::io::detail;
+
+  auto writer = make_writer<csv::writer>(args.sink(), args, mr);
+
+  writer->write_all(args.table(), args.metadata());
+}
+
 namespace orc = cudf::experimental::io::detail::orc;
 
 // Freeform API wraps the detail reader class API